feedback delay example
